second commit

+++ database/migrations/2021_03_05_203447_create_wallets_table.php
@@ -18,7 +18,7 @@ class CreateWalletsTable extends Migration
             $table->string('name');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('currency_id');
-            $table->integer('balance')->default(0);
+            $table->decimal('balance', 10, 2)->default(0);
 
             $table->timestamps();
